Trading_System/requirements.txt
# =====================================================
# Trading System Requirements
# Python 3.10+
# =====================================================

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Technical Analysis
TA-Lib>=0.4.28
pandas-ta>=0.3.14b

# Interactive Brokers API
ib_insync>=0.9.86

# Visualization
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Framework (for Dashboard)
fastapi>=0.100.0
uvicorn>=0.22.0
websockets>=11.0

# Configuration Management
PyYAML>=6.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Logging & Monitoring
loguru>=0.7.0
python-json-logger>=2.0.7

# Notifications
python-telegram-bot>=20.3
requests>=2.31.0

# Async Support
asyncio>=3.4.3
aiohttp>=3.8.4

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
tqdm>=4.65.0

# Performance
numba>=0.57.0
cython>=0.29.35

# Machine Learning (Optional - for AI features)
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Development Tools
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
ipython>=8.14.0
jupyter>=1.0.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0
